Fixed performing migrations to Sqlite database

+++ database/migrations/main/2021_09_10_115126_create_cashier_increment_field.php
@@ -27,14 +27,18 @@ class CreateCashierIncrementField extends Migration
     public function up()
     {
         Schema::table($this->table(), function (Blueprint $table) {
-            $table->bigIncrements('id')->first();
+            if ($this->doesntMemory()) {
+                $table->bigIncrements('id')->first();
+            }
         });
     }
 
     public function down()
     {
         Schema::table($this->table(), function (Blueprint $table) {
-            $table->dropColumn('id');
+            if ($this->doesntMemory()) {
+                $table->dropColumn('id');
+            }
         });
     }
